+++ src/rdkafka_request.c
@@ -507,9 +507,9 @@ void rd_kafka_OffsetFetchRequest (rd_kafka_broker_t *rkb,
                                   rd_kafka_resp_cb_t *resp_cb,
                                   void *opaque) {
 	rd_kafka_buf_t *rkbuf;
-        int of_TopicCnt;
+        size_t of_TopicCnt;
         int TopicCnt = 0;
-        int of_PartCnt = -1;
+        ssize_t of_PartCnt = -1;
         const char *last_topic = NULL;
         int PartCnt = 0;
         int i;
@@ -659,10 +659,10 @@ void rd_kafka_OffsetCommitRequest (rd_kafka_broker_t *rkb,
                                    rd_kafka_resp_cb_t *resp_cb,
                                    void *opaque) {
 	rd_kafka_buf_t *rkbuf;
-        int of_TopicCnt = -1;
+        ssize_t of_TopicCnt = -1;
         int TopicCnt = 0;
         const char *last_topic = NULL;
-        int of_PartCnt = -1;
+        ssize_t of_PartCnt = -1;
         int PartCnt = 0;
         int i;
 
@@ -758,8 +758,8 @@ static rd_kafkap_bytes_t *rd_kafka_group_MemberState_consumer_write (
         rd_kafka_buf_t *rkbuf;
         int i;
         const char *last_topic = NULL;
-        int of_TopicCnt;
-        int of_PartCnt = -1;
+        size_t of_TopicCnt;
+        ssize_t of_PartCnt = -1;
         int TopicCnt = 0;
         int PartCnt = 0;
         rd_kafkap_bytes_t *MemberState;
